[AIRFLOW-1889] Add sensors in sensors.py to operators package
Adds sensors from sensors.py to the operators package, removing the
sensors.py file altogether. Maintains old import style (to be removed in
2.0).
